修改README，并对定时器做修改

+++ src/main/java/cn/dombro/dataimport/service/TaskService.java
@@ -72,13 +72,17 @@ class DeleteTask extends Task{
             return;
         }else {
             File[] childrenFiles = dir.listFiles();
-            for (int i = 0; i < childrenFiles.length; i++){
-                 if (childrenFiles[i].isFile()){
-                     FilePathEnum.DELETE_FILE.deleteFile(childrenFiles[i].getAbsolutePath());
-                 } else {
-                     deleteDir(childrenFiles[i]);
-                     childrenFiles[i].delete();
-                 }
+            if (childrenFiles.length == 0) {
+                for (int i = 0; i < childrenFiles.length; i++) {
+                    if (childrenFiles[i].isFile()) {
+                        FilePathEnum.DELETE_FILE.deleteFile(childrenFiles[i].getAbsolutePath());
+                    } else {
+                        deleteDir(childrenFiles[i]);
+                        childrenFiles[i].delete();
+                    }
+                }
+            }else {
+                return;
             }
 
         }
